posting-function 画像の投稿後にstateの初期化を行うように変更しました。
ImageInputのファイル名をCreatePostに変更しました。
firebaseにアップロード後の初期化が実現できなかったため、投稿画面に導入していたreduxを削除しました。
