+++ src/app/store.ts
@@ -1,11 +1,9 @@
-import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import userReducer from '../features/userSlice';
-import postReducer from '../features/postSlice';
+import { configureStore, ThunkAction, Action } from "@reduxjs/toolkit";
+import userReducer from "../features/userSlice";
 
 export const store = configureStore({
   reducer: {
     user: userReducer,
-    post: postReducer,
   },
 });
